requirements.txt
apache-airflow-providers-amazon

psycopg2-binary
requests
python-dotenv

streamlit
boto3
pandas

numpy<2
pyarrow
fastparquet
snowflake-connector-python==3.13.2

openpyxl
pyxlsb
xlrd
odfpy
dbfread

PyMuPDF
python-docx
lxml

duckdb
SQLAlchemy

geopandas
shapely
pyproj
pyogrio
affine
rasterio==1.4.2

ezdxf
py7zr
rarfile

paramiko